requirements.txt
typeannotations==0.1.0
bitarray==0.8.1
nose==1.3.7
wheel==0.24.0
